+++ packages/agents/lighthouse/src/lib/operations/index.ts
@@ -1,9 +1,11 @@
 import { execute } from "./execute";
 import { sendToRelayer } from "./relayer";
+import { pollCartographer } from "./cartographer";
 
 export const getOperations = () => {
   return {
     execute,
     sendToRelayer,
+    pollCartographer,
   };
 };
