+++ packages/agents/lighthouse/test/mock.ts
@@ -168,6 +168,7 @@ export const mock = {
   operations: {
     execute: stub(),
     sendToRelayer: stub(),
+    pollCartographer: stub(),
   },
 };
